Implemented Leapfrog algorithm

+++ Project5/Solvers.h
@@ -13,9 +13,9 @@ protected:
 	SolarSystem* _rk4;
 	SolarSystem* _leapfrog;
 	SolarSystem* _euler;
-	void RK4(int step);
-	void Leapfrog(int step);
-	void Euler(int step);
+	void RK4(double step);
+	void Leapfrog(double step);
+	void Euler(double step);
 public:
 	Solvers(SolarSystem* system, bool useRK4, bool useLeapfrog, bool useEuler);
 	~Solvers();
